% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo_10.R
\name{algo_10}
\alias{algo_10}
\title{ALGO_10 function}
\usage{
algo_10(input, input2, name_users)
}
\arguments{
\item{input}{a dfm avec lignes/users et col/text}

\item{input2}{a df with users and target (df validation)}

\item{name_users}{the name of col users in the df validation (in "")}
}
\value{
data frame with the validation score of the 10 sample vector calib and
}
\description{
find the best sample for calibration
}
\examples{
w<-data_fixure()
dfm<-w[[1]]
val<-w[[2]]
remove(w)
s<-algo_10(dfm,val,"users_id")
}
