% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo_10_review.R
\name{algo_10_review}
\alias{algo_10_review}
\title{ALGO_10_REVIEW function}
\usage{
algo_10_review(input, input2, name_users)
}
\arguments{
\item{input}{a dfm avec lignes/users et col/text}

\item{input2}{a df with users and target (df validation)}

\item{name_users}{the name of col users in the df validation (in "")}
}
\value{
data frame with the validation score of the 10 sample vector calib and
}
\description{
find the best sample for calibration
}
\examples{
w<-data_fixture()
algo_10_review(w[[1]],w[[2]],"users_id")
remove(w)
}
