% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate}
\alias{calibrate}
\title{Calibrate function}
\usage{
calibrate(
  input,
  complet = T,
  calib,
  ngram = 2,
  wordsinrow = FALSE,
  docincol = FALSE
)
}
\arguments{
\item{complet}{}
}
\value{

}
\description{
Calibrate the weight of words to applicate on the users
}
