% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fixture.R
\name{data_fixture}
\alias{data_fixture}
\title{Make the data fixture with the df validation}
\usage{
data_fixture()
}
\value{

}
\description{
Make the data fixture with the df validation
}
\examples{
w <- data_fixture()
dfm <- w[[1]]
val <- w[[2]]
}
