% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm_generation.R
\name{dfm_generation}
\alias{dfm_generation}
\title{dfm_generation}
\usage{
dfm_generation(input, text, doc, lg)
}
\arguments{
\item{input}{un data frame avec au moins une colonne représentant l'ID
de l'user et une colonne représentant le text/Tweet}

\item{text}{le numéro de la colonne du text/Tweet}

\item{doc}{le  numéro de la colonne ID}

\item{lg}{le language des Tweets (une suele langue)}
}
\value{
la fonction renvoie le dfm
}
\description{
Création d'une Documents-Features-Matrix à partir d'un data frame de
texte et d'users
}
