% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x2y.R
\name{dx2y}
\alias{dx2y}
\title{Multiple metrics x2y}
\usage{
dx2y(d, target = NA, confidence = FALSE)
}
\arguments{
\item{d}{f}

\item{target}{f}

\item{confidence}{f}
}
\value{

}
\description{
Multiple metrics x2y
}
