% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_prediction.R
\name{Estimate_prediction}
\alias{Estimate_prediction}
\title{Estimate and predict function}
\usage{
Estimate_prediction(
  input,
  ID,
  calib,
  ngram = 2,
  wordsinrow = FALSE,
  docincol = FALSE
)
}
\arguments{
\item{docincol}{}
}
\value{

}
\description{
Give de estimation of the target with calculcing the weight of words
}
\examples{
source("data/Ex_Tweets.csv")
Tweet<-read.csv('Ex_Tweets.csv',encoding = 'UTF-8')
Tweet<-subset(Tweet, select = c("user_id","text"))
DFM_test<-dfm_generation(Tweet,2,1,"fr")
List_test<-Estimate_prediction(DFM_test,"user_id",c(1,2,3))
}
