% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_fixure.R
\name{make_data_fixure}
\alias{make_data_fixure}
\title{Make a Data fixure}
\usage{
make_data_fixure(nb_users)
}
\arguments{
\item{nb_users}{}
}
\value{

}
\description{
Make a Data fixure
}
\examples{
Data_test<-make_data_fixure(1000)

}
