% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_random_data_fixture.R
\name{make_random_data_fixture}
\alias{make_random_data_fixture}
\title{Make a random Data fixture}
\usage{
make_random_data_fixture(nb_users)
}
\arguments{
\item{nb_users}{the numbers of users you want}
}
\value{

}
\description{
Make a random Data fixture
}
\examples{
make_random_data_fixture(1000)
}
