% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_random_data_fixure.R
\name{make_random_data_fixure}
\alias{make_random_data_fixure}
\title{Make a random Data fixure}
\usage{
make_random_data_fixure(nb_users)
}
\arguments{
\item{nb_users}{the numbers of users you want}
}
\value{

}
\description{
Make a random Data fixure
}
\examples{
Data_test<-make_data_fixure(1000)
}
