% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_weight.R
\name{use_weight}
\alias{use_weight}
\title{Use weight function}
\usage{
use_weight(input, words, df, wordsinrow = FALSE, docincol = FALSE)
}
\arguments{
\item{docincol}{}
}
\value{

}
\description{
Prediction on the users
}
