% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validation}
\alias{validation}
\title{Validation function}
\usage{
validation(input_1, input_2, name_user_1, name_user_2)
}
\arguments{
\item{input_1}{the data frame with the score calculate}

\item{input_2}{the real score}

\item{name_user_1}{the name of the col of the users-id in the df calculate}

\item{name_user_2}{the name of the col of the users-id in the df validation}
}
\value{

}
\description{
Validation function
}
\examples{
 w <- data_fixture()

 x<-calibrate(w[[1]],complet=T,c(1,2,3))

 y <- use_weight(x[[2]],rownames(x[[1]]),x[[1]])

 opinions_df <- rbind(x[[3]], y[[1]])
 opinions_df$users <- as.numeric(opinions_df$users)
 opinions_df$opinions <- as.numeric(opinions_df$opinions)

 validation(opinions_df,w[[2]],"users","users_id")
 remove(w)
 remove(x)
 remove(y)
}
