% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordfish.R
\name{wordfish}
\alias{wordfish}
\title{Title}
\usage{
wordfish(
  input,
  wordsincol = FALSE,
  fixtwo = FALSE,
  dir = NULL,
  fixdoc = c(1, 2, 0, 1),
  tol = 1e-07,
  sigma = 3,
  boots = FALSE,
  nsim = 500,
  writeout = FALSE,
  output = "wordfish_output"
)
}
\arguments{
\item{output}{}
}
\value{

}
\description{
Title
}
