% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate}
\alias{calibrate}
\title{Calibrate function}
\usage{
calibrate(
  input,
  complet = T,
  calib,
  ngram = 2,
  wordsinrow = FALSE,
  docincol = FALSE
)
}
\arguments{
\item{input}{the dfm}

\item{complet}{T or F if the dfm is complet}

\item{calib}{the vector calib}

\item{ngram}{fix}

\item{wordsinrow}{fix}

\item{docincol}{fix}
}
\value{

}
\description{
Calibrate the weight of words to applicate on the users
}
\examples{
w <- data_fixture()
dfm <- w[[1]]
calibrate(dfm,complet=T,c(1,2,3))
}
