% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_weight.R
\name{use_weight}
\alias{use_weight}
\title{Use weight function}
\usage{
use_weight(input, words, df, wordsinrow = FALSE, docincol = FALSE)
}
\arguments{
\item{input}{a data.frame with the users avec lignes/users et col/text}

\item{df}{word/weight data calculate in the calibrate function}

\item{wordsinrow}{fix}

\item{docincol}{fix}

\item{word}{words kept calculating in the calibrate function}
}
\value{

}
\description{
Prediction on the users
}
\examples{
w <- data_fixure()
dfm <- w[[1]]
calibrate(dfm,complet=T,c(1,2,3))
  data_users <- x[[2]]
  word_df <- x[[1]]
  opini_target <- x[[3]]
y <- use_weight(data_users,rownames(word_df),word_df)
  opini_users <- y[[1]]
  word_wei <- y[[2]]
}
