% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_weight.R
\name{use_weight}
\alias{use_weight}
\title{Use weight function}
\usage{
use_weight(input, words, df, wordsinrow = FALSE, docincol = FALSE)
}
\arguments{
\item{input}{a data.frame with the users avec lignes/users et col/text}

\item{words}{words kept calculating in the calibrate function}

\item{df}{word/weight data calculate in the calibrate function}

\item{wordsinrow}{fix}

\item{docincol}{fix}
}
\value{

}
\description{
Prediction on the users
}
\examples{
w <- data_fixture()
x<-calibrate(w[[1]],complet=T,c(1,2,3))
use_weight(x[[2]],rownames(x[[1]]),x[[1]])
remove(w)
remove(x)
}
