% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validation}
\alias{validation}
\title{Validation functioon}
\usage{
validation(input_1, input_2)
}
\arguments{
\item{input_1}{the data frame with the score calculate}

\item{input_2}{the real score}
}
\value{

}
\description{
Validation functioon
}
